pyproject.toml
[project]
name = "dnd_dm_copilot"
version = "0.1.0"
description = "A context-aware RAG system for D&D Dungeon Masters - fine-tuned embedding model for campaign assistance"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # LLM Interface
    "openai",
    # Core ML libraries
    "datasets<4.0",
    "sentence-transformers>=5.1.0",
    "transformers[torch]>=4.40.0",
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    # Data processing
    "pandas>=2.0.0",
    "tqdm>=4.65.0",
    "huggingface-hub>=0.20.0",
    # Visualization
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "plotly>=5.15.0",
    "umap-learn>=0.5.0",
    # API and HTTP
    "requests>=2.31.0",
    "httpx>=0.24.0",
    # FastAPI backend
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    # PDF processing
    "pypdf>=3.17.0",
    # Vector store
    "faiss-cpu>=1.7.0",
    # Configuration and utilities
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
    # Development and testing
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.1",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "dotenv>=0.9.9",
    "jsonlines>=4.0.0",
    "wandb>=0.22.2",
]

[project.optional-dependencies]
llm = [
    "llama-cpp-python>=0.2.0",
]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = [
    "datasets",
    "datasets.*",
    "sklearn.*",
    "wandb",
    "wandb.*",
    "faiss",
    "pypdf",
    "llama_cpp",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=dnd_dm_copilot --cov-report=term-missing --cov-report=html --cov-report=xml --strict-markers"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
branch = true
source = ["dnd_dm_copilot"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__main__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
exclude = [".git", "__pycache__", ".venv", "venv"]
max-complexity = 10
